This is the merge of the ARM64 feature work into the official/rs5 branch.  This results in one branch for both architectures.

Related work items: #15331345

+++ ArmPkg/Drivers/ArmGic/GicV2/ArmGicV2SecLib.c
@@ -65,7 +65,7 @@ ArmGicV2SetupNonSecure (
   }
 
   // Ensure all interrupts can get through the priority mask
-  MmioWrite32 (GicInterruptInterfaceBase + ARM_GIC_ICCPMR, CachedPriorityMask);
+  MmioWrite32 (GicInterruptInterfaceBase + ARM_GIC_ICCPMR, (UINT32)CachedPriorityMask);
 }
 
 VOID
